Update register view, app view, and Import Velocity Dashboard Style.css
We updated the register view with the appropriate fields required to successfully register an account. Modification where also made to the app view in order to change fonts and also pull the velocity dashboard stylesheet. This is not the original version of the stylesheet, as it was modified to suit this project. It has a few *minor* with additions of styles. We also added a new folders "images/auth" in the public root and added the register view background. To learn more about the victory dashboard use the following link:

https://themeforest.net/item/victory-bootstrap-4-admin-template/21223393