Allow login via email/username, change login view, and add image
We updated the LoginController to allow both authentication using email/username. We also update the login view to be able to accept both email/username under the same input called "identity". We change the directory names for where the application images and logos are placed. Uploaded image for login page and update the style.css to use that image. Added the BTC Logo to login screen.