Created Overview Page and Pointed (Login, Registration, Password Reset and Authentication) to Overview Page
Created the overview page where the user will always land if authenticated. Also created the overview route to connect to view. Middleware added for login page to avoid user accessing the page when logged in as well as middleware for overview page so that user cannont access when logged out. The Login, Registration, Password Reset Controllers and Authentication Middleware are pointed to always move the user towards the overview page on successful requests.